LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/reporterror-cross-realm-method-expected.txt

CONSOLE MESSAGE: TypeError: foo


PASS self.reportError() dispatches an "error" event for this's relevant global object
